changes to rice crop.

+++ src/main/java/net/tarksmods/ricemod/blocks/RiceCropBlock.java
@@ -9,6 +9,8 @@ import net.tarksmods.ricemod.items.ModItems;
 
 public class RiceCropBlock extends CropBlock {
     private static final VoxelShape[] AGE_TO_SHAPE = new VoxelShape[]{
+
+            //This determines the hitbox of each growth stage.
             Block.createCuboidShape(0.0D, 0.0D, 0.0D, 16.0D, 2.0D, 16.0D),
             Block.createCuboidShape(0.0D, 0.0D, 0.0D, 16.0D, 3.0D, 16.0D),
             Block.createCuboidShape(0.0D, 0.0D, 0.0D, 16.0D, 4.0D, 16.0D),
@@ -17,14 +19,6 @@ public class RiceCropBlock extends CropBlock {
             Block.createCuboidShape(0.0D, 0.0D, 0.0D, 16.0D, 7.0D, 16.0D),
             Block.createCuboidShape(0.0D, 0.0D, 0.0D, 16.0D, 8.0D, 16.0D),
             Block.createCuboidShape(0.0D, 0.0D, 0.0D, 16.0D, 9.0D, 16.0D)
-//            Block.createCuboidShape(0.0D, 0.0D, 0.0D, 0.0D, 2.0D, 0.0D), //0
-//            Block.createCuboidShape(0.0D, 0.0D, 0.0D, 0.0D, 3.0D, 0.0D), //1
-//            Block.createCuboidShape(0.0D, 0.0D, 0.0D, 0.0D, 4.0D, 0.0D), //2
-//            Block.createCuboidShape(0.0D, 0.0D, 0.0D, 0.0D, 5.0D, 0.0D), //3
-//            Block.createCuboidShape(0.0D, 0.0D, 0.0D, 0.0D, 6.0D, 0.0D), //4
-//            Block.createCuboidShape(0.0D, 0.0D, 0.0D, 0.0D, 7.0D, 0.0D), //5
-//            Block.createCuboidShape(0.0D, 0.0D, 0.0D, 0.0D, 8.0D, 0.0D), //6
-//            Block.createCuboidShape(0.0D, 0.0D, 0.0D, 0.0D, 9.0D, 0.0D)  //7
     };
 
     public RiceCropBlock(AbstractBlock.Settings settings) {
@@ -35,6 +29,6 @@ public class RiceCropBlock extends CropBlock {
         return ModItems.RICE_ITEM;
     }
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        return AGE_TO_SHAPE[(Integer) state.get(this.getAgeProperty())];
+        return AGE_TO_SHAPE[state.get(this.getAgeProperty())];
     }
 }
